// src/app/components/shared/image-upload/interfaces.ts
export interface ImageResult {
  file: File;
  url: string;
  dataURL?: string;
  resized?: {
    dataURL: string;
    type: string;
    width: number;
    height: number;
  };
}
export interface ResizeOptions {
  resizeMaxHeight?: number;
  resizeMaxWidth?: number;
  resizeQuality?: number;
  resizeType?: string;
}